+++ xnt-core/src/protocol/consensus/transaction/validity/kernel_to_outputs.rs
@@ -536,6 +536,6 @@ mod tests {
 
     test_program_snapshot!(
         KernelToOutputs,
-        "cbcebd905a23309a96f22c5acf57e691cb796078c7b88c2ca50bef35ad74cb3718163a593ca79f26"
+        "74e12c483cfd1c57a76bfaafa2ea2b92177103367b9cdc0ab2bd3491c051599fbbaef1b77cc00c75"
     );
 }